+++ src/Language/Futhark/TypeChecker/Monad.hs
@@ -17,7 +17,6 @@ module Language.Futhark.TypeChecker.Monad
   , localEnv
 
   , TypeError(..)
-  , prettyTypeError, prettyTypeErrorNoLoc
   , unexpectedType
   , undefinedType
   , unappliedFunctor
@@ -89,8 +88,7 @@ instance Pretty Note where
   ppr (Note msg) = "Note:" <+> align msg
 
 instance Pretty Notes where
-  ppr (Notes notes) = folddoc twolines $ map ppr notes
-    where twolines x y = x <> line <> line <> y
+  ppr (Notes notes) = foldMap (((line<>line)<>) . ppr) notes
 
 aNote :: Pretty a => a -> Notes
 aNote = Notes . pure . Note . ppr
@@ -98,13 +96,10 @@ aNote = Notes . pure . Note . ppr
 -- | Information about an error during type checking.
 data TypeError = TypeError SrcLoc Notes Doc
 
-prettyTypeErrorNoLoc :: TypeError -> String
-prettyTypeErrorNoLoc (TypeError _ notes msg) =
-  pretty $ msg <> ppr notes
-
-prettyTypeError :: TypeError -> String
-prettyTypeError e@(TypeError loc _ _) =
-  "Error at " ++ locStr loc ++ ":\n" ++ prettyTypeErrorNoLoc e
+instance Pretty TypeError where
+  ppr (TypeError loc notes msg) =
+    "Error at" <+> text (locStr loc) <+> ":" </>
+    msg <> ppr notes
 
 unexpectedType :: MonadTypeChecker m => SrcLoc -> StructType -> [StructType] -> m a
 unexpectedType loc _ [] =